+++ CHANGELOG.md
@@ -8,6 +8,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - Find static methods on enums #737
 - Support completion after using try operator `?` #726
 - Find methods on cooked string literals #728
+- Find doc comments on named and indexed struct fields #739
 
 ## 2.0.8